% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numint.R
\name{int}
\alias{int}
\title{Auxiliary univariate numerical integration}
\usage{
int(f, x, tol = 1e-10)
}
\arguments{
\item{f}{Vector of values of function f at x}

\item{x}{Equispaced and odd length vector}

\item{tol}{Error tolerance}
}
\value{
Scalar value of numerical integration of function f over x
}
\description{
Auxiliary univariate numerical integration
}
