% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture_density.R
\name{mixture_density}
\alias{mixture_density}
\title{Mixture density estimate}
\usage{
mixture_density(f, U, d, Xsup, ...)
}
\arguments{
\item{f}{Vector of mixing density values or vector of particle approximation over support U}

\item{U}{Mixing density support, where U is a
\itemize{
\item Equispaced and odd length vector for univariate f
\item Matrix of two columns, where each column is the support for the corresponding variate for bivariate f
\item A matrix where each row is a random sample from f0
}}

\item{d}{Parametric kernel function of form d(x,u)}

\item{Xsup}{Vector or matrix (each row representing one support point) at which mixture density is calculated}
}
\value{
Vector containing mixture density estimate values at Xsup
}
\description{
Calculates the mixture density at a user-defined grid of points given a mixing density
}
\seealso{
\code{\link[=pr]{pr()}}
}
