% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monotone.R
\name{monotone}
\alias{monotone}
\title{Monotone density estimation using PR}
\usage{
monotone(X, nperm = 25)
}
\arguments{
\item{X}{Vector of data values}

\item{nperm}{Number of permutations for the PR estimate}
}
\value{
A list with the following elements,
\itemize{
\item m - Estimated monotone density
\item Xsup - Corresponding support
}
}
\description{
Estimates a monotone density via the PR algorithm
}
\references{
Dixit, Vaidehi, and Ryan Martin. "Revisiting consistency of a
recursive estimator of mixing distributions."
arXiv preprint arXiv:2110.02465 (2021).
}
