% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prreg.R
\name{prem.reg}
\alias{prem.reg}
\title{Hybrid predictive recursion-EM algorithm for regression}
\source{
A first version of the code is available at \url{https://www4.stat.ncsu.edu/~rmartin/Codes/prreg.R}
}
\usage{
prem.reg(X, Y, perm, get.ci = FALSE)
}
\arguments{
\item{X}{Matrix of covariate values}

\item{Y}{Vector of response values}

\item{perm}{A matrix of user-defined permutation sequence containing \code{nperm} rows and number of columns equal to number of data points in X}

\item{get.ci}{Logical. Should confidence intervals be calculated for the coefficients? Default is \link{FALSE}}
}
\value{
A list with the following elements,
\itemize{
\item b - Estimated regression coefficients
\item Yhat - Predicted responses at X
\item ci - Confidence intervals for b
\item f - Permutation-averaged mixing density estimate
\item UU - Vector of support for f
\item wt - Final vector of weights in the regression fit
\item lik - Permutation-averaged log-likelihood function L
}
}
\description{
Fits a robust linear regression model
}
\examples{
# Belgian phone call data in Example 1 of the paper
phone.example <- function() {
require(MASS)
data(phones)
Y <- phones$calls
X <- phones$year + 1900
o <- prem.reg(X=cbind(1 + 0 * Y, X), Y=Y)
plot(x=X, y=Y, xlab="Year", ylab="Calls (in millions)")
abline(a=o$b[1], b=o$b[2])
abline(lm(Y ~ X), lty=2)
legend(x="topleft", inset=0.05, lty=1:2, c("PREM", "LS"))
return(list(b=o$b, f=o$f, UU=o$UU, w=o$wt, lik=o$lik))
}

#TO RUN
phone.example()
}
\references{
Martin, Ryan, and Zhen Han. "A semiparametric scale-mixture regression model
and predictive recursion maximum likelihood."
Computational Statistics & Data Analysis 94 (2016): 75-85.
}
\seealso{
\code{\link[=pr]{pr()}}
}
