% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numint.R
\name{simp.int2}
\alias{simp.int2}
\title{Auxiliary bivariate numerical integration}
\usage{
simp.int2(x, y, Fxy, tol = 1e-10)
}
\arguments{
\item{x}{Equispaced vector}

\item{y}{Equispaced vector}

\item{Fxy}{Matrix of dimension dim(x) * dim(y) that contains values of f(x,y)}

\item{tol}{Error tolerance}
}
\value{
Scalar value of numerical integration of function f over x
}
\description{
Auxiliary bivariate numerical integration
}
