% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture_density.R
\name{single.mix.multi}
\alias{single.mix.multi}
\title{Mixture density at point x for multidimensional U}
\usage{
single.mix.multi(x, D, U, d, t, ...)
}
\arguments{
\item{x}{Single value at which the mixture density is to be calculated}

\item{D}{Mixing density weights at U}

\item{U}{See documentation for mixture_density}

\item{d}{See documentation for mixture_density}

\item{t}{Number of rows of U}
}
\value{
Mixture density calculated at x
}
\description{
Mixture density at point x for multidimensional U
}
\seealso{
\code{\link[=pr]{pr()}}, \code{\link[=mixture_density]{mixture_density()}}
}
