% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture_density.R
\name{single.mix1}
\alias{single.mix1}
\title{Mixture density at point x for unidimensional U}
\usage{
single.mix1(x, f, U, d, ...)
}
\arguments{
\item{x}{Single value at which the mixture density is to be calculated}

\item{f}{See documentation for mixture_density}

\item{U}{See documentation for mixture_density}

\item{d}{See documentation for mixture_density}
}
\value{
Mixture density calculated at x
}
\description{
Mixture density at point x for unidimensional U
}
\seealso{
\code{\link[=pr]{pr()}}, \code{\link[=mixture_density]{mixture_density()}}
}
