% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monotone.R
\name{monotone}
\alias{monotone}
\title{Monotone density estimation using PR}
\usage{
monotone(X, nperm = 25)
}
\arguments{
\item{X}{Vector of data values}

\item{nperm}{Number of permutations for the PR estimate}
}
\value{
A list with the following elements,
\itemize{
\item m - Estimated monotone density
\item Xsup - Corresponding support
}
}
\description{
Estimates a monotone density via the PR algorithm
}
\examples{
# To estimate a monotone density using PR
# Generate data
n = 200
X = abs(rnorm(n))
# Estimate
m = monotone(X)
# Plot with the truth
truth = 2*dnorm(m$Xsup)
plot(m)
lines(m$Xsup, m$m, col = "red")
}
\references{
Dixit, Vaidehi, and Ryan Martin. "Revisiting consistency of a
recursive estimator of mixing distributions."
arXiv preprint arXiv:2110.02465 (2021).
}
